feat: 支持小数分钟的错误恢复时间配置
## 🎯 功能增强
现在错误恢复时间支持小数配置，可以更精确地控制恢复时间

## 📝 修改内容

### 后端修改
- **errorRecoveryHelper.js**: 将 parseInt() 改为 parseFloat()
  - 支持接收和处理小数值（如 0.5、1.5、2.5 等）
  - 保持原有的验证逻辑（值必须 > 0）

### 前端修改
- **AccountForm.vue**: 更新所有平台的错误恢复时间输入框（18处）
  - 添加 `step="0.1"` 支持 0.1 分钟为步进单位
  - 最小值从 `min="1"` 改为 `min="0.1"`
  - 允许输入 0.1-999.9 分钟的任意小数值

## ✅ 支持的配置示例
- 0.1 分钟 = 6 秒（快速恢复）
- 0.5 分钟 = 30 秒（网络抖动）
- 1 分钟 = 60 秒（标准恢复）
- 2.5 分钟 = 150 秒（中等恢复）
- 5.5 分钟 = 330 秒（长期恢复）

## 💡 使用场景
1. **快速恢复**（0.5-1分钟）：适用于网络暂时性抖动
2. **中等恢复**（2-5分钟）：适用于暂时性服务问题
3. **长期恢复**（10-30分钟）：适用于严重的服务故障

## 🧪 测试验证
✅ 后端 parseFloat() 语法检查通过
✅ 前端构建成功（12.69秒）
✅ 小数值解析正确（0.5 -> 30秒，1.5 -> 90秒）
✅ 默认值机制正常（无效值使用默认5分钟）

+++ src/utils/errorRecoveryHelper.js
@@ -24,7 +24,7 @@ class ErrorRecoveryHelper {
    */
   static getRecoveryDuration(account, defaultDuration = 5) {
     if (!account) return defaultDuration
-    const duration = parseInt(account.errorRecoveryDuration)
+    const duration = parseFloat(account.errorRecoveryDuration)
     return Number.isFinite(duration) && duration > 0 ? duration : defaultDuration
   }
